MAINT: move generate_instrument_list to utils.core
