+++ pysat/tests/instrument_test_class.py
@@ -45,87 +45,6 @@ def remove_files(inst):
                                'ensure temp directory is used')))
 
 
-def generate_instrument_list(inst_loc):
-    """Iterate through and create all of the test Instruments needed.
-
-
-    Parameters
-    ----------
-    inst_loc : python subpackage
-        The location of the instrument subpackage to test,
-        eg, 'pysat.instruments'
-
-    Note
-    ----
-    - Only want to do this once per instrument subpackage being tested.
-    - There are a number of checks here to skip import issues for certain cases.
-      These are caught later in the tests below as part of InstTestClass. This
-      is done to ensure the success of this routine so that tests for all
-      instruments can be run uninterupted even if one instrument is broken.
-
-    """
-
-    instrument_names = inst_loc.__all__
-    instrument_download = []
-    instrument_no_download = []
-
-    # create temporary directory
-    dir_name = tempfile.mkdtemp()
-    saved_path = pysat.data_dir
-    pysat.utils.set_data_dir(dir_name, store=False)
-
-    # Look through list of available instrument modules in the given location
-    for inst_module in instrument_names:
-        try:
-            module = import_module(''.join(('.', inst_module)),
-                                   package=inst_loc.__name__)
-        except ImportError:
-            # If this can't be imported, we can't pull out the info for the
-            # download / no_download tests.  Leaving in basic tests for all
-            # instruments, but skipping the rest.  The import error will be
-            # caught as part of the pytest.mark.all_inst tests in InstTestClass
-            pass
-        else:
-            # try to grab basic information about the module so we
-            # can iterate over all of the options
-            try:
-                info = module._test_dates
-            except AttributeError:
-                # If a module does not have a test date, add it anyway for
-                # other tests.  This will be caught later by
-                # InstTestClass.test_instrument_test_dates
-                info = {}
-                info[''] = {'': dt.datetime(2009, 1, 1)}
-                module._test_dates = info
-            for sat_id in info.keys():
-                for tag in info[sat_id].keys():
-                    inst_dict = {'inst_module': module, 'tag': tag,
-                                 'sat_id': sat_id}
-                    # Initialize instrument so that pysat can generate skip
-                    # flags where appropriate
-                    inst = pysat.Instrument(inst_module=module,
-                                            tag=tag,
-                                            sat_id=sat_id,
-                                            temporary_file_list=True)
-                    travis_skip = ((os.environ.get('TRAVIS') == 'true')
-                                   and not inst._test_download_travis)
-                    if inst._test_download:
-                        if not travis_skip:
-                            instrument_download.append(inst_dict)
-                    elif not inst._password_req:
-                        # we don't want to test download for this combo
-                        # But we do want to test the download warnings
-                        # for instruments without a password requirement
-                        instrument_no_download.append(inst_dict)
-    pysat.utils.set_data_dir(saved_path, store=False)
-
-    output = {'names': instrument_names,
-              'download': instrument_download,
-              'no_download': instrument_no_download}
-
-    return output
-
-
 def initialize_test_inst_and_date(inst_dict):
     """Initializes the instrument object to test and date
