+++ pysat/tests/test_instruments.py
@@ -3,13 +3,13 @@ import pytest
 # Make sure to import your instrument package here
 import pysat
 # Import the test classes from pysat
-from pysat.tests.instrument_test_class import generate_instrument_list
+from pysat.utils import generate_instrument_list
 from pysat.tests.instrument_test_class import InstTestClass
 
 # Developers for instrument libraries should update the following line to
 # point to their own subpackage location
 # e.g.,
-# instruments = generate_instrument_list(inst_loc=mypackage.instruments)
+# instruments = generate_instrument_list(inst_loc=mypackage.inst)
 instruments = generate_instrument_list(inst_loc=pysat.instruments)
 
 # The following lines apply the custom instrument lists to each type of test
@@ -39,6 +39,11 @@ class TestInstruments(InstTestClass):
 
     def setup(self):
         """Runs before every method to create a clean testing setup."""
+        # Make sure to use a temporary directory so that the user's setup is not
+        # altered
+        self.tempdir = tempfile.TemporaryDirectory()
+        self.saved_path = pysat.data_dir
+        pysat.utils.set_data_dir(self.tempdir.name, store=False)
         # Developers for instrument libraries should update the following line
         # to point to their own subpackage location, e.g.,
         # self.inst_loc = mypackage.instruments
